requirements.txt
fastmcp
python-dotenv
splitwise